Added CPU Usage function

+++ cpu/cpu_usage.h
@@ -7,6 +7,6 @@
 #ifndef CPU_USAGE_H
 #define CPU_USAGE_H
 
-int get_cpu_usage(void);
+double get_cpu_usage(void);
 
 #endif
